+++ js/util.js
@@ -1,0 +1,29 @@
+function getRandomPositiveInteger(a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+//Функция для проверки максимальной длины строки
+
+function getLength(description, length) {
+  return description.length <= length;
+}
+
+getLength('Текстовый комментарий', 140);
+
+//массив случайной длины из значений, значения не повторяются
+function getListElements(counter) {
+  const newArray = [];
+  for (let i = 1; i <= counter; i++) {
+    const newElem = i;
+    newArray.push(newElem);
+  }
+  return newArray;
+}
+
+//даёт одно значение из массива
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveInteger, getListElements, getRandomArrayElement};
